Clean up frametime / delta time calculations. Use Q16.16 numbers for better accuracy. Fix bug with rotation (delta_y was being calculated from delta_x that had been rotated already).
Accuracy of rotation is still no great probably because of fixedptc.h's inaccurate sine/cosine. With a large rotation angle like -15 º, my mouse cursor starts slowly climbing up the screen with horizontal mouse swiping. The best way to fix this might be to go back to libfixmath now that the frametime calculations are fixed up for Q16.16.